都営地下鉄の利用者数の推移: Rename components/MetroBarChart.vue.vue to components/index/CardsReference/Metro/Chart.vue
